# Conflicts:
#	src/main/java/at/ac/tuwien/inso/service/impl/CourseServiceImpl.java
#	src/main/java/at/ac/tuwien/inso/service/impl/FeedbackServiceImpl.java
#	src/main/java/at/ac/tuwien/inso/service/impl/StudyPlanServiceImpl.java
#	src/main/java/at/ac/tuwien/inso/service/impl/SubjectServiceImpl.java
#	src/main/java/at/ac/tuwien/inso/service/impl/UisUserServiceImpl.java
#	src/main/java/at/ac/tuwien/inso/service/impl/UserCreationServiceImpl.java

+++ src/main/java/at/ac/tuwien/inso/service/impl/SubjectServiceImpl.java
@@ -30,9 +30,9 @@ import at.ac.tuwien.inso.service.SubjectService;
 @Service
 public class SubjectServiceImpl implements SubjectService {
 
-    private static final Logger logger = LoggerFactory.getLogger(SubjectServiceImpl.class);
     private ValidatorFactory validatorFactory = new ValidatorFactory();
     private SubjectValidator validator = validatorFactory.getSubjectValidator();
+    private static final Logger log = LoggerFactory.getLogger(SubjectServiceImpl.class);
 
     @Autowired
     private SubjectRepository subjectRepository;
@@ -45,6 +45,7 @@ public class SubjectServiceImpl implements SubjectService {
 
     @Override
     public Page<Subject> findBySearch(String search, Pageable pageable) {
+    	log.info("finding search by word "+search);
         String sqlSearch = "%" + search + "%";
         return subjectRepository.findAllByNameLikeIgnoreCase(sqlSearch, pageable);
     }
@@ -53,10 +54,11 @@ public class SubjectServiceImpl implements SubjectService {
     @Transactional(readOnly = true)
     public Subject findOne(Long id) {
         validator.validateSubjectId(id);
+    	log.info("finding subject by id "+id);
         Subject subject = subjectRepository.findOne(id);
 
         if(subject == null) {
-            logger.warn("Subject not found");
+            log.warn("Subject not found");
             //TODO throwing this will cause a security test fail, for whatever reason?
             //throw new SubjectNotFoundException();
         }
@@ -68,13 +70,14 @@ public class SubjectServiceImpl implements SubjectService {
     @Transactional
     public Subject create(Subject subject) {
         validator.validateNewSubject(subject);
+    	log.info("creating subject "+subject.toString());
         return subjectRepository.save(subject);
     }
 
     @Override
     @Transactional
     public Lecturer addLecturerToSubject(Long subjectId, Long lecturerUisUserId) {
-        logger.info("addLecturerToSubject for subject {} and lecturer {}", subjectId,
+        log.info("addLecturerToSubject for subject {} and lecturer {}", subjectId,
                 lecturerUisUserId);
 
         validator.validateSubjectId(subjectId);
@@ -82,7 +85,7 @@ public class SubjectServiceImpl implements SubjectService {
 
         if (lecturer == null) {
             String msg = "Lecturer with user id " + lecturerUisUserId + " not found";
-            logger.info(msg);
+            log.info(msg);
             throw new LecturerNotFoundException(msg);
         }
 
@@ -90,7 +93,7 @@ public class SubjectServiceImpl implements SubjectService {
 
         if (subject == null) {
             String msg = "Subject with id " + lecturerUisUserId + " not found";
-            logger.info(msg);
+            log.info(msg);
             throw new SubjectNotFoundException(msg);
         }
 
@@ -112,10 +115,13 @@ public class SubjectServiceImpl implements SubjectService {
         if (search == null) {
             search = "";
         }
-
+    	log.info("getting available lectureres for subject with subject id "+subjectId+" and search string "+search);
+        
+    	
         Subject subject = subjectRepository.findById(subjectId);
 
         if (subject == null) {
+        	log.warn("Subject with id '" + subjectId + "' not found");
             throw new SubjectNotFoundException("Subject with id '" + subjectId + "' not found");
         }
 
@@ -137,14 +143,13 @@ public class SubjectServiceImpl implements SubjectService {
     @Override
     public Lecturer removeLecturerFromSubject(Long subjectId, Long lecturerUisUserId) {
         validator.validateSubjectId(subjectId);
-        logger.info("removeLecturerFromSubject for subject {} and lecturer {}", subjectId,
-                lecturerUisUserId);
+        log.info("removeLecturerFromSubject for subject {} and lecturer {}", subjectId, lecturerUisUserId);
 
         Subject subject = subjectRepository.findById(subjectId);
 
         if (subject == null) {
             String msg = "Subject with id '" + subjectId + "' not found";
-            logger.info(msg);
+            log.info(msg);
             throw new SubjectNotFoundException(msg);
         }
 
@@ -152,7 +157,7 @@ public class SubjectServiceImpl implements SubjectService {
 
         if (lecturer == null) {
             String msg = "Lecturer with id '" + lecturerUisUserId + "' not found";
-            logger.info(msg);
+            log.info(msg);
             throw new LecturerNotFoundException(msg);
         }
 
@@ -163,7 +168,7 @@ public class SubjectServiceImpl implements SubjectService {
         if (!isLecturer) {
             String msg = "Lecturer with id " + lecturerUisUserId + " not found for subject " +
                     subjectId;
-            logger.info(msg);
+            log.info(msg);
             throw new RelationNotfoundException(msg);
         }
 
@@ -177,20 +182,27 @@ public class SubjectServiceImpl implements SubjectService {
     @Override
     @Transactional(readOnly = true)
     public List<Subject> searchForSubjects(String word) {
+    	log.info("seachring for subjects with search word "+word);
         return subjectRepository.findByNameContainingIgnoreCase(word);
     }
 
 	@Override
 	@Transactional
 	public boolean remove(Subject subject) throws ValidationException{
-        validator.validateNewSubject(subject);
+		log.info("trying to remove subject");
+		validator.validateNewSubject(subject);
+		if(subject==null){
+			log.info("Subject is null.");
+			throw new ValidationException("Subject is null.");
+		}
+		log.info("removing subject "+subject+" now.");
 		List<Course> courses = courseService.findCoursesForSubject(subject);
 		if(!courses.isEmpty()){
 			String msg = "";
 			if(subject!=null){
 				msg = "Cannot delete subject [Name: "+subject.getName()+"] because there are courses";
 			}
-			logger.info(msg);
+			log.info(msg);
 			throw new ValidationException(msg);
 		}else{
 			subjectRepository.delete(subject);
